third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-play-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
Tests normal playback case with MediaSource API

EVENT(webkitsourceopen)
EVENT(webkitsourceended)
EVENT(playing)
EVENT(ended)
END OF TEST
